CAMEL-15088: camel-dapr - refactor consumer
